Add files via upload
As part of a college assignment, I had to design an app for "AWGN Channel Performance Analysis". I used the kivy framework for python to describe the frontend of the app. The input is a number taken from the user, and the results are visually shown on a semilog graph.